fetcher/requirements.txt
aio-pika==9.5.5
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiormq==6.8.1
aiosignal==1.3.2
alibabacloud-darabonba-string==0.0.4
alibabacloud-kms20160120==2.2.3
alibabacloud-tea==0.4.3
alibabacloud_credentials==0.3.6
alibabacloud_darabonba_array==0.1.0
alibabacloud_darabonba_encode_util==0.0.2
alibabacloud_darabonba_map==0.0.1
alibabacloud_darabonba_signature_util==0.0.4
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_pop==0.0.8
alibabacloud_gateway_spi==0.0.3
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.13
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.7.1
cryptography==44.0.2
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.0
multidict==6.2.0
nacos-sdk-python==2.0.3
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.2
pamqp==3.3.0
playwright==1.51.0
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.27.2
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.0
starlette==0.46.1
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
python-consul2==0.1.5
